+++ resources/JSON/gameInteraction.json
