+++ resources/META-INF/MANIFEST.MF
