+++ src/com/Group3/ZombieBytes/Client/ZombieGameClient.java
@@ -1,18 +1,6 @@
 package com.Group3.ZombieBytes.Client;
 
 import com.Group3.ZombieBytes.Game.Game;
-import com.Group3.ZombieBytes.PrintToOutPut.PrintContent;
-//import com.Group3.ZombieBytes.lifeforms.Character;
-//import com.Group3.ZombieBytes.lifeforms.Zombie;
-import com.Group3.ZombieBytes.Game.GameText;
-import com.Group3.ZombieBytes.Items.*;
-import org.json.simple.*;
-import com.Group3.ZombieBytes.Game.Location;
-import org.json.simple.parser.*;
-
-import java.io.FileReader;
-import java.io.IOException;
-import java.util.ArrayList;
 
 public class ZombieGameClient {
     public static void main(String[] args) {
