+++ src/com/Group3/ZombieBytes/Data/Location.java
@@ -1,11 +1,10 @@
-package com.Group3.ZombieBytes.Game;
+package com.Group3.ZombieBytes.Data;
 
-import com.Group3.ZombieBytes.lifeforms.Zombie;
-import com.Group3.ZombieBytes.Items.Item;
+import com.Group3.ZombieBytes.Data.Lifeforms.Zombie;
+import com.Group3.ZombieBytes.Data.Items.Item;
 
 import java.util.ArrayList;
 import java.util.HashMap;
-import java.util.List;
 
 public class Location {
     // properties
