+++ src/com/Group3/ZombieBytes/Util/JsonParser/ZombieParser.java
@@ -1,9 +1,8 @@
-package com.Group3.ZombieBytes.JsonParser;
+package com.Group3.ZombieBytes.Util.JsonParser;
 
 import com.Group3.ZombieBytes.Game.Game;
-import com.Group3.ZombieBytes.Game.Location;
-import com.Group3.ZombieBytes.Items.Item;
-import com.Group3.ZombieBytes.lifeforms.Zombie;
+import com.Group3.ZombieBytes.Data.Location;
+import com.Group3.ZombieBytes.Data.Lifeforms.Zombie;
 import org.json.simple.JSONArray;
 import org.json.simple.JSONObject;
 import org.json.simple.parser.JSONParser;
